# templates/trykatch/web/packages/api-client/openapi/Trykatch.Api.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Trykatch API",
    "description": "The versioned HTTP contract for Trykatch's platform and organization workspaces. Organization-scoped operations infer their organization from the authenticated workspace context.",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/account/security/reauthenticate": {
      "post": {
        "tags": [
          "AccountSecurity"
        ],
        "summary": "Reauthenticate account security",
        "operationId": "AccountSecurity_Reauthenticate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReauthenticationRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ReauthenticationRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ReauthenticationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RecentAssuranceGrant"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecentAssuranceGrant"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecentAssuranceGrant"
                }
              }
            }
          }
        },
        "x-codeSamples": [
          {
            "lang": "TypeScript",
            "label": "TypeScript · fetch",
            "source": "const { token } = await fetch('/api/v1/auth/antiforgery', {\n  credentials: 'include',\n}).then((response) => response.json() as Promise<{ token: string }>);\n\nconst response = await fetch('/api/v1/account/security/reauthenticate', {\n  method: 'POST',\n  credentials: 'include',\n  headers: {\n    Accept: 'application/json',\n    'Content-Type': 'application/json',\n    'X-CSRF-TOKEN': token,\n  },\n  body: JSON.stringify({ /* fields from the request schema */ }),\n});\n\nif (!response.ok) throw new Error(`Request failed: ${response.status}`);\nconst result: unknown = response.status === 204 ? undefined : await response.json();\n// Handle result in memory; never log response bodies, which may contain credentials or personal data.\nconsole.log({ status: response.status });"
          }
        ]
      }
    },
    "/api/v1/account/security/mfa/setup": {
      "post": {
        "tags": [
          "AccountSecurity"
        ],
        "summary": "Setup mfa account security",
        "operationId": "AccountSecurity_SetupMfa",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaGrantRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaGrantRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MfaGrantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PendingMfaSetup"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingMfaSetup"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingMfaSetup"
                }
              }
            }
          }
        },
        "x-codeSamples": [
          {
            "lang": "TypeScript",
            "label": "TypeScript · fetch",
            "source": "const { token } = await fetch('/api/v1/auth/antiforgery', {\n  credentials: 'include',\n}).then((response) => response.json() as Promise<{ token: string }>);\n\nconst response = await fetch('/api/v1/account/security/mfa/setup', {\n  method: 'POST',\n  credentials: 'include',\n  headers: {\n    Accept: 'application/json',\n    'Content-Type': 'application/json',\n    'X-CSRF-TOKEN': token,\n  },\n  body: JSON.stringify({ /* fields from the request schema */ }),\n});\n\nif (!response.ok) throw new Error(`Request failed: ${response.status}`);\nconst result: unknown = response.status === 204 ? undefined : await response.json();\n// Handle result in memory; never log response bodies, which may contain credentials or personal data.\nconsole.log({ status: response.status });"
          }
        ]
      }
    },
    "/api/v1/account/security/mfa/enable": {
      "post": {
        "tags": [
          "AccountSecurity"
        ],
        "summary": "Enable mfa account security",
        "operationId": "AccountSecurity_EnableMfa",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaCodeRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaCodeRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MfaCodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MfaRecoveryCodes"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MfaRecoveryCodes"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MfaRecoveryCodes"
                }
              }
            }
          }
        },
        "x-codeSamples": [
          {
            "lang": "TypeScript",
            "label": "TypeScript · fetch",
            "source": "const { token } = await fetch('/api/v1/auth/antiforgery', {\n  credentials: 'include',\n}).then((response) => response.json() as Promise<{ token: string }>);\n\nconst response = await fetch('/api/v1/account/security/mfa/enable', {\n  method: 'POST',\n  credentials: 'include',\n  headers: {\n    Accept: 'application/json',\n    'Content-Type': 'application/json',\n    'X-CSRF-TOKEN': token,\n  },\n  body: JSON.stringify({ /* fields from the request schema */ }),\n});\n\nif (!response.ok) throw new Error(`Request failed: ${response.status}`);\nconst result: unknown = response.status === 204 ? undefined : await response.json();\n// Handle result in memory; never log response bodies, which may contain credentials or personal data.\nconsole.log({ status: response.status });"
          }
        ]
      }
    },
    "/api/v1/account/security/mfa/cancel": {
      "post": {
        "tags": [
          "AccountSecurity"
        ],
        "summary": "Cancel mfa enrollment account security",
        "operationId": "AccountSecurity_CancelMfaEnrollment",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaEnrollmentRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaEnrollmentRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MfaEnrollmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "x-codeSamples": [
          {
            "lang": "TypeScript",
            "label": "TypeScript · fetch",
            "source": "const { token } = await fetch('/api/v1/auth/antiforgery', {\n  credentials: 'include',\n}).then((response) => response.json() as Promise<{ token: string }>);\n\nconst response = await fetch('/api/v1/account/security/mfa/cancel', {\n  method: 'POST',\n  credentials: 'include',\n  headers: {\n    Accept: 'application/json',\n    'Content-Type': 'application/json',\n    'X-CSRF-TOKEN': token,\n  },\n  body: JSON.stringify({ /* fields from the request schema */ }),\n});\n\nif (!response.ok) throw new Error(`Request failed: ${response.status}`);\nconst result: unknown = response.status === 204 ? undefined : await response.json();\n// Handle result in memory; never log response bodies, which may contain credentials or personal data.\nconsole.log({ status: response.status });"
          }
        ]
      }
    },
    "/api/v1/account/security/mfa/recovery-codes": {
      "post": {
        "tags": [
          "AccountSecurity"
        ],
        "summary": "Regenerate recovery codes account security",
        "operationId": "AccountSecurity_RegenerateRecoveryCodes",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaGrantRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaGrantRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MfaGrantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MfaRecoveryCodes"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MfaRecoveryCodes"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MfaRecoveryCodes"
                }
              }
            }
          }
        },
        "x-codeSamples": [
          {
            "lang": "TypeScript",
            "label": "TypeScript · fetch",
            "source": "const { token } = await fetch('/api/v1/auth/antiforgery', {\n  credentials: 'include',\n}).then((response) => response.json() as Promise<{ token: string }>);\n\nconst response = await fetch('/api/v1/account/security/mfa/recovery-codes', {\n  method: 'POST',\n  credentials: 'include',\n  headers: {\n    Accept: 'application/json',\n    'Content-Type': 'application/json',\n    'X-CSRF-TOKEN': token,\n  },\n  body: JSON.stringify({ /* fields from the request schema */ }),\n});\n\nif (!response.ok) throw new Error(`Request failed: ${response.status}`);\nconst result: unknown = response.status === 204 ? undefined : await response.json();\n// Handle result in memory; never log response bodies, which may contain credentials or personal data.\nconsole.log({ status: response.status });"
          }
        ]
      }
    },
    "/api/v1/account/security/mfa/disable": {
      "post": {
        "tags": [
          "AccountSecurity"
        ],
        "summary": "Disable mfa account security",
        "operationId": "AccountSecurity_DisableMfa",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaGrantRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaGrantRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MfaGrantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "x-codeSamples": [
          {
            "lang": "TypeScript",
            "label": "TypeScript · fetch",
            "source": "const { token } = await fetch('/api/v1/auth/antiforgery', {\n  credentials: 'include',\n}).then((response) => response.json() as Promise<{ token: string }>);\n\nconst response = await fetch('/api/v1/account/security/mfa/disable', {\n  method: 'POST',\n  credentials: 'include',\n  headers: {\n    Accept: 'application/json',\n    'Content-Type': 'application/json',\n    'X-CSRF-TOKEN': token,\n  },\n  body: JSON.stringify({ /* fields from the request schema */ }),\n});\n\nif (!response.ok) throw new Error(`Request failed: ${response.status}`);\nconst result: unknown = response.status === 204 ? undefined : await response.json();\n// Handle result in memory; never log response bodies, which may contain credentials or personal data.\nconsole.log({ status: response.status });"
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MfaCodeRequest": {
        "required": [
          "enrollmentId",
          "code"
        ],
        "type": "object",
        "properties": {
          "enrollmentId": {
            "type": "string",
            "format": "uuid"
          },
          "code": {
            "type": "string"
          }
        }
      },
      "MfaEnrollmentRequest": {
        "required": [
          "enrollmentId"
        ],
        "type": "object",
        "properties": {
          "enrollmentId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "MfaGrantRequest": {
        "type": "object",
        "properties": {
          "grant": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "MfaRecoveryCodes": {
        "required": [
          "codes"
        ],
        "type": "object",
        "properties": {
          "codes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "signInRequired": {
            "type": "boolean",
            "default": true
          }
        }
      },
...
      "PendingMfaSetup": {
        "required": [
          "enrollmentId",
          "sharedKey",
          "authenticatorUri",
          "expiresAt"
        ],
        "type": "object",
        "properties": {
          "enrollmentId": {
            "type": "string",
            "format": "uuid"
          },
          "sharedKey": {
            "type": "string"
          },
          "authenticatorUri": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ReauthenticationRequest": {
        "required": [
          "purpose",
          "password"
        ],
        "type": "object",
        "properties": {
          "purpose": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "code": {
            "type": [
              "null",
              "string"
            ]
          },
          "isRecoveryCode": {
            "type": "boolean",
            "default": false
          }
        }
      },
      "RecentAssuranceGrant": {
        "required": [
          "grant",
          "expiresAt"
        ],
        "type": "object",
        "properties": {
          "grant": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
...
}