# templates/trykatch/web/packages/api-client/src/generated/models/mfaGrantRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Trykatch API
 * The versioned HTTP contract for Trykatch's platform and organization workspaces. Organization-scoped operations infer their organization from the authenticated workspace context.
 * OpenAPI spec version: v1
 */

export interface MfaGrantRequest {
  /** @nullable */
  grant?: string | null;
}